feat: add Sanity schema types for author, category, post, block content, and SEO; refactor imports for consistency
