+++ src/sanity/schema-types/category-type.ts
