+++ src/sanity/schema-types/seo-type.ts
